requirements.txt
streamlit==1.38.0
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
psycopg2-binary==2.9.9
Pillow==10.1.0
streamlit-option-menu==0.4.0
python-dotenv=1.0.1